+++ lib/database/migrations/2017_02_21_155940_create_facility_table.php
@@ -14,7 +14,7 @@ class CreateFacilityTable extends Migration
     public function up()
     {
         Schema::create('tbl_facility', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->integer('code');
             $table->string('postal_address');
             $table->string('email_address');
@@ -24,9 +24,9 @@ class CreateFacilityTable extends Migration
             $table->integer('weekend_max');
             $table->integer('weekday_max');
             $table->tinyInteger('is_sms');
-            $table->integer('county_sub_id')->unsigned();
-            $table->integer('supporter_id')->unsigned();
-            $table->integer('facility_type_id')->unsigned();
+            $table->bigInteger('county_sub_id')->unsigned();
+            $table->bigInteger('supporter_id')->unsigned();
+            $table->bigInteger('facility_type_id')->unsigned();
             $table->timestamps();
             $table->softDeletes();
