+++ lib/database/migrations/2017_02_23_140854_create_regimen_drug_table.php
@@ -14,11 +14,11 @@ class CreateRegimenDrugTable extends Migration
     public function up()
     {
         Schema::create('tbl_regimen_drug', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->integer('ccc_store_sp');
             $table->string('source');
-            $table->integer('drug_id')->unsigned(); 
-            $table->integer('regimen_id')->unsigned();  
+            $table->bigInteger('drug_id')->unsigned(); 
+            $table->bigInteger('regimen_id')->unsigned();  
             $table->timestamps();
             $table->softDeletes();
